requirements.txt
flask
python-telegram-bot==13.15
requests
pytz
ta
pandas
apscheduler
